turn likelihood module into a package
